Edit-V1
FINISHED:
- Extracting Components to a separated folder for reusable
- Authentication

IN PROGRESS:
- Storage
