+++ README.md
@@ -1,2 +1,11 @@
 # Consensus2024
  We SUOHA!!!
+
+
+## Fun Facts that are useful (JS)
+1. Function Declarations cannot be reassigned.
+
+## PROGRESS
+- [ ] Seperate All Components For Future Purposes
+- [ ] Website Layouts
+- [ ]
