+++ frontend/libs/storage.ts
@@ -1,58 +1,44 @@
-import { userSession } from "./auth";
-import { Storage } from '@stacks/storage';
+// Initiating a storage client
+
+"use server"
+
+import { hasAppPrivateKey } from "@stacks/connect";
+import { getUserKeys, userSession, getUserData, getPerson } from "./userSession"
+import { Storage } from "@stacks/storage";
+
+// User's PrivateKey 
+const { userPrivateKey, userPublicKey } = getUserKeys();
+
+// Getting App PrivateKey (the appPrivateKey is automatically derived and provided by the authentication process)
+const privateKey = userSession.loadUserData().appPrivateKey
+
+const storage = new Storage({userSession});
 
 
-export const storage = new Storage({ userSession });
 
-const PROFILE_FILENAME = 'profile.json';
 
-export const saveProfile = async (profile: any) => {
-  return await storage.putFile(
-    PROFILE_FILENAME,
-    JSON.stringify({ profile, isPublic: true }),
-    {
-      encrypt: false,
-      dangerouslyIgnoreEtag: true
-    }
-  );
-};
 
-export const fetchProfile = async (username: any) => {
-  try {
-    const profileJson: any = await storage.getFile(PROFILE_FILENAME, {
-      decrypt: false,
-      username: username || undefined,
-      app: 'https://Eleutheria.com',
-      verify: false
-    });
 
-    if (profileJson) {
-      const json = JSON.parse(profileJson);
-      return {
-        profile: json?.profile
-      };
-    }
-    return {
-      profile: null
-    };
-  } catch (e) {
-    console.log('refetch with old domain https://Eleutheria.com');
-    const profileJson: any = await storage.getFile(PROFILE_FILENAME, {
-      decrypt: false,
-      username: username || undefined,
-      app: 'https://Eleutheria.com',
-      verify: false
-    });
 
-    if (profileJson) {
-      const json = JSON.parse(profileJson);
-      console.log('profileJson', json);
-      return {
-        profile: json?.profile
-      };
-    }
-    return {
-      profile: null
-    };
-  }
-};
+
+
+
+
+
+
+
+
+
+// TESTING:
+
+// Storing AppPrivateKey into the UserSession
+// userSession.store.getSessionData().userData = <any>{
+//     appPrivateKey: privateKey,
+// };
+
+// userSession.store.getSessionData().userData = <any> {
+//     hasAppPrivateKey
+// }
+
+
+
