+++ frontend/libs/v1/auth.ts
