+++ frontend/package.json
@@ -12,6 +12,7 @@
     "@splinetool/runtime": "^1.4.0",
     "@stacks/auth": "^6.15.0",
     "@stacks/connect": "^7.7.1",
+    "@stacks/encryption": "^6.15.0",
     "@stacks/network": "^6.13.0",
     "@stacks/storage": "^6.15.0",
     "@stacks/transactions": "^6.15.0",
